+++ tests/api/test_checkout.py
@@ -6,7 +6,7 @@ import pytest
 
 from saleor.checkout.models import Cart
 from saleor.checkout.utils import (
-    add_voucher_to_cart, is_fully_paid, ready_to_place_order)
+    add_voucher_to_cart, can_be_fully_paid, ready_to_place_order)
 from saleor.graphql.core.utils import str_to_enum
 from saleor.order.models import Order
 from tests.api.utils import get_graphql_content
@@ -802,6 +802,7 @@ def test_checkout_complete(
     checkout = cart_with_item
     checkout.shipping_address = address
     checkout.shipping_method = shipping_method
+    checkout.billing_address = address
     checkout.save()
 
     checkout_line = checkout.lines.first()
@@ -867,6 +868,7 @@ def test_checkout_complete_no_payment(
     checkout = cart_with_item
     checkout.shipping_address = address
     checkout.shipping_method = shipping_method
+    checkout.billing_address = address
     checkout.save()
     checkout_id = graphene.Node.to_global_id('Checkout', checkout.pk)
     variables = {'checkoutId': checkout_id}
@@ -875,7 +877,8 @@ def test_checkout_complete_no_payment(
         MUTATION_CHECKOUT_COMPLETE, variables)
     content = get_graphql_content(response)
     data = content['data']['checkoutComplete']
-    assert data['errors'][0]['message'] == 'Checkout is not fully paid'
+    assert data['errors'][0]['message'] == \
+        'Checkout can not be fully paid with proper payments'
     assert orders_count == Order.objects.count()
 
 
@@ -889,6 +892,7 @@ def test_checkout_complete_insufficient_stock(
     cart_line.save()
     checkout.shipping_address = address
     checkout.shipping_method = shipping_method
+    checkout.billing_address = address
     checkout.save()
     total = checkout.get_total()
     payment = payment_dummy
@@ -1092,6 +1096,7 @@ def test_ready_to_place_order(
     checkout = cart_with_item
     checkout.shipping_address = address
     checkout.shipping_method = shipping_method
+    checkout.billing_address = address
     checkout.save()
     total = checkout.get_total()
     payment = payment_dummy
@@ -1142,13 +1147,14 @@ def test_ready_to_place_order_no_payment(
     checkout = cart_with_item
     checkout.shipping_address = address
     checkout.shipping_method = shipping_method
+    checkout.billing_address = address
     checkout.save()
     ready, error = ready_to_place_order(checkout, None, None)
     assert not ready
-    assert error == 'Checkout is not fully paid'
+    assert error == 'Checkout can not be fully paid with proper payments'
 
 
-def test_is_fully_paid(cart_with_item, payment_dummy):
+def test_can_be_fully_paid(cart_with_item, payment_dummy):
     checkout = cart_with_item
     total = checkout.get_total()
     payment = payment_dummy
@@ -1158,11 +1164,11 @@ def test_is_fully_paid(cart_with_item, payment_dummy):
     payment.currency = total.gross.currency
     payment.checkout = checkout
     payment.save()
-    is_paid = is_fully_paid(checkout, None, None)
+    is_paid = can_be_fully_paid(checkout, None, None)
     assert is_paid
 
 
-def test_is_fully_paid_many_payments(cart_with_item, payment_dummy):
+def test_can_be_fully_paid_many_payments(cart_with_item, payment_dummy):
     checkout = cart_with_item
     total = checkout.get_total()
     payment = payment_dummy
@@ -1180,11 +1186,11 @@ def test_is_fully_paid_many_payments(cart_with_item, payment_dummy):
     payment2.currency = total.gross.currency
     payment2.checkout = checkout
     payment2.save()
-    is_paid = is_fully_paid(checkout, None, None)
+    is_paid = can_be_fully_paid(checkout, None, None)
     assert is_paid
 
 
-def test_is_fully_paid_partially_paid(cart_with_item, payment_dummy):
+def test_can_be_fully_paid_partially_paid(cart_with_item, payment_dummy):
     checkout = cart_with_item
     total = checkout.get_total()
     payment = payment_dummy
@@ -1194,13 +1200,13 @@ def test_is_fully_paid_partially_paid(cart_with_item, payment_dummy):
     payment.currency = total.gross.currency
     payment.checkout = checkout
     payment.save()
-    is_paid = is_fully_paid(checkout, None, None)
+    is_paid = can_be_fully_paid(checkout, None, None)
     assert not is_paid
 
 
-def test_is_fully_paid_no_payment(cart_with_item):
+def test_can_be_fully_paid_no_payment(cart_with_item):
     checkout = cart_with_item
-    is_paid = is_fully_paid(checkout, None, None)
+    is_paid = can_be_fully_paid(checkout, None, None)
     assert not is_paid